requirements.txt
# Web框架和API
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
email-validator>=2.0.0
pydantic>=2.0.0
pydantic-core>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0
starlette>=0.30.0

# 数据库和ORM
sqlalchemy>=2.0.0
alembic>=1.10.0
pymysql>=1.0.3
psycopg2-binary>=2.9.5
redis>=4.5.4
SQLAlchemy-Utils>=0.41.0

# 认证和安全
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
pyjwt>=2.6.0
python-dotenv>=1.0.0

# 文件存储
minio>=7.1.13
aiofiles>=23.1.0
Pillow>=9.5.0

# 邮件
fastapi-mail>=1.2.8

# AI和ML集成
openai>=1.0.0
anthropic>=0.4.0
langchain>=0.0.267
langchain-core>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.0.10
langsmith>=0.0.57
sentence-transformers>=2.2.2
deepseek-ai>=0.0.1
tiktoken>=0.4.0

# 科学计算和数据处理
scikit-learn>=1.2.2
numpy>=1.24.2
pandas>=2.0.0
matplotlib>=3.7.1

# 文档处理
pypdf>=3.7.0
PyMuPDF>=1.21.1
markdown>=3.4.3
python-docx>=0.8.11
python-pptx>=0.6.21
pdfminer.six>=20221105
fpdf>=1.7.2
reportlab>=4.0.4

# 文本和NLP
nltk>=3.8.1
spacy>=3.6.0
jieba>=0.42.1
langdetect>=1.0.9
transformers>=4.30.0
tokenizers>=0.13.3

# 论文搜索和分析
arxiv>=2.0.0
scholarly>=1.7.11
bibtexparser>=1.4.0
pybtex>=0.24.0
pyalex>=0.10
xmltodict>=0.13.0
lxml>=4.9.2
beautifulsoup4>=4.11.2
crossref-commons>=0.0.8

# 网络请求和抓取
requests>=2.31.0
aiohttp>=3.8.4
fake-useragent>=1.1.3
free-proxy>=1.1.0
selenium>=4.10.0

# 工具和辅助库
pyyaml>=6.0
tqdm>=4.65.0
tenacity>=8.2.2
loguru>=0.7.0
typer>=0.9.0
click>=8.1.3
rich>=13.4.2

# 异步和并发
asyncio>=3.4.3
celery>=5.2.7
APScheduler>=3.10.1

# 格式化和验证
jsonschema>=4.17.3
pydocstyle>=6.3.0
black>=23.3.0
isort>=5.12.0

# 测试和开发工具
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
coverage>=7.2.7

# Web前端集成
flask>=2.3.2
werkzeug>=2.3.6
jinja2>=3.1.2
gunicorn>=21.2.0
waitress>=2.1.2

# 监控和调试
prometheus-client>=0.17.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0

# 跨域支持
fastapi-cors>=0.0.6

# 系统和环境
psutil>=5.9.5
python-multipart>=0.0.6

# 缓存和性能
cachetools>=5.3.1
chromadb>=0.4.13
faiss-cpu>=1.7.4 